Simplify codebase and workflows
