+++ .github/workflows/ci.yaml
@@ -27,9 +27,15 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python }}
+      - name: Generate horreum client
+        run: make generate
       - name: Install poetry
         run: |
           pipx install --pip-args=--constraint=./dev-constraints.txt poetry
           poetry --version
+      - name: Install dependencies
+        run: poetry install
       - name: Build python library
         run: poetry build
+      - name: Test horreum library
+        run: poetry run pytest test/horreum_client_test.py
