--- .github/workflows/update-client.yaml
@@ -1,48 +1,0 @@
-# This workflow will fetch the new openapi from the provided branch of https://github.com/Hyperfoil/Horreum and based on
-# that it will re-generate the Horreum raw client and creates a new pull request against the corresponding branch here.
-# It could be tested running `gh act workflow_dispatch -e ./test/workflow_dispatch_event_example.json`. Remember to
-# comment out the pull request creation :)
-name: Update Horreum auto-generated client
-
-on:
-  workflow_dispatch:
-    inputs:
-      branch:
-        description: Branch or tag of https://github.com/Hyperfoil/Horreum
-        required: true
-  repository_dispatch:
-    types: [ generate-horreum-client ]
-
-jobs:
-  generate:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Fetch branch
-        id: fetch-branch
-        run: |
-          if [ "${{ github.event_name }}" = "repository_dispatch" ]; then
-            # Generated by peter-evans/repository-dispatch@v3
-            echo "HORREUM_BRANCH=${{ github.event.client_payload.branch }}" >> $GITHUB_OUTPUT
-          elif [ "${{ github.event_name }}" = "workflow_dispatch" ]; then
-            echo "HORREUM_BRANCH=${{ github.event.inputs.branch }}" >> $GITHUB_OUTPUT
-          else
-            echo "Unknown event: ${{ github.event_name }}"
-            exit 1
-          fi
-      - uses: actions/checkout@v4
-        with:
-          persist-credentials: false
-      - name: Generate horreum client
-        run: make HORREUM_BRANCH=${{ steps.fetch-branch.outputs.HORREUM_BRANCH }} generate
-      - run: git --no-pager diff
-      - name: Create Pull Request
-        uses: gr2m/create-or-update-pull-request-action@v1
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        with:
-          title: Generate Horreum client from github.com/Hyperfoil/Horreum:${{ steps.fetch-branch.outputs.HORREUM_BRANCH }}
-          body: |
-            There is a new change in the openapi spec of Horreum in branch ${{ steps.fetch-branch.outputs.HORREUM_BRANCH }}.
-            Verify that there are no breaking changes.
-          branch: horreum-openapi-${{ steps.fetch-branch.outputs.HORREUM_BRANCH }}
-          commit-message: Horreum openapi client updated
