+++ pyproject.toml
@@ -16,7 +16,7 @@ include = [
 "Issues" = "https://github.com/Hyperfoil/horreum-client-python/issues"
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = "^3.9"
 microsoft-kiota-abstractions = "^1.3.1"
 microsoft-kiota-http = "^1.3.1"
 microsoft-kiota-serialization-json = "^1.1.0"
