--- src/horreum/service/__init__.py
@@ -1,5 +1,0 @@
-# from .info_service import get_client_version
-
-__all__ = [
-    "info_service"
-]
