--- src/horreum/service/info_service.py
@@ -1,5 +1,0 @@
-from importlib.metadata import version
-
-
-def get_client_version():
-    return version("horreum")
